Rewrite playbook to be several smaller imported playbooks

+++ bootstrap.sh
@@ -28,30 +28,72 @@ EOF
     exit
 }
 
+# A prompt function
+function ask() {
+    local ans
+    while true; do
+        printf "\n%s [y|N] ? " "$@"
+        read -r ans
+        case "$ans" in
+        y* | Y*) return 0 ;;
+        *) return 1 ;;
+        esac
+    done
+}
 
-# Install curl, uv, initialize a project
-apt install -y curl
-curl -LsSf https://astral.sh/uv/install.sh | env UV_INSTALL_DIR="/tmp/uv" sh
-source /tmp/uv/env
-mkdir -p /tmp/bootstrap
-pushd /tmp/bootstrap &>/dev/null || exit 1
-uv run --no-project ansible \
-	-c "ansible-galaxy collection install -f git+https://github.com/markfaine/net-markfaine.git"
+function prompt(){
+    local prompt ans default
+    prompt="${1:-None}"
+    default="${2:-''}"
+    while true; do
+        printf "\n%s: [%s]" "$prompt" "$default"
+        read -r ans
+        if [[ "${ans:-}" == "" ]]; then
+            break
+        fi
+        if ask "You entered '$ans', is this correct"; then
+            break
+        fi
+    done
+    echo "$ans"
+}
 
-if ! uv run --no-project ansible \
-	-c "ansible-playbook ~/.ansible/collections/ansible_collections/net/markfaine/playbooks/playbook.yml; then
-    printf "Sorry there was an error!\n"
-    exit 1
+
+# Install curl
+if ! command -v curl &>/dev/null; then
+    printf "Installing curl\n"
+    apt install -y curl
 fi
 
-cat <<EOF 
-  Login as the target user and run the followng commands:
+# Install uv
+printf "Installing and configuring uv\n"
+curl -LsSf https://astral.sh/uv/install.sh | env UV_INSTALL_DIR="/tmp/uv" sh &>/dev/null
+mkdir -p /tmp/bootstrap
+pushd /tmp/bootstrap &>/dev/null || exit 1
 
-  # If you have a dotenv
-  # npx dotenv-vault@latest pull [ENVIRONMENT]
+printf "Installing collection\n"
+uvx --from ansible-core ansible-galaxy collection install -f git+https://github.com/markfaine/net-markfaine.git,experimental &>/dev/null
 
-  # Install/Update mise packages
-  # mise up # by default this will also pull tuckr dotfiles from the specified dotfiles repo
-  
+# Set base command
+playbook_cmd=(uvx --from ansible-core ansible-playbook)
 
+# Set target host
+hostname="$(prompt "Target Hostname" "localhost")"
+playbook_cmd+=(-l "$hostname")
+if [[ "${hostname:-}" == "localhost" ]]; then
+    playbook_cmd+=(--connection local)
+fi
 
+roles=(packages user mise wsl docker fonts)
+run_roles=()
+for role in "${roles[@]}"; do
+    if ask "Run the $role role"; then
+        run_roles+=($role) 
+    fi
+done
+tags="$(IFS=','; echo "${run_roles[*]}")"
+playbook_cmd+=(-t "$tags")
+playbook_cmd+=(~/.ansible/collections/ansible_collections/net/markfaine/playbooks/playbook.yml)
+
+printf "Run playbooks\n"
+"${playbook_cmd[@]}"
